Struct light_cli::LightCliInput

pub struct LightCliInput<SLEN> where
    SLEN: ArrayLength<u8>,  { /* fields omitted */ }

Methods

impl<SLEN: ArrayLength<u8>> LightCliInput<SLEN>

pub fn new() -> Self

Create a new LightCLI instance.

pub fn parse_data<CB>(&mut self, callback: CB) -> Result<(), Error> where
CB: FnMut(CallbackCommand),

Try to parse as much data from the internal ring buffer as possible.

Arguments

callback - This is the callback that will receive all parsing events.

Remarks

All commands are in the form "COMMAND KEY=VALUE". For every parsed key/value pair the callback will be triggered with the current command string, current key and the corresponding value. When a newline is read the callback is triggered with a command event.

pub fn fill<E>(&mut self, ser: &mut dyn Read<u8, Error = E>) -> Result<(), E>

Copy as many available bytes from ser into the buffer as possible.

Arguments

ser - The serial interface to read from.

Remarks

This will continue to try to read a byte from the serial device until the device returns nb::Error::WouldBlock.